third_party/blink/web_tests/platform/generic/svg/dynamic-updates/SVGFEDiffuseLightingElement-dom-lighting-color-attr-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'lighting-color' attribute of the SVGFEDiffuseLightingElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS gradientElement.getAttribute('lighting-color') is "green"
PASS gradientElement.getAttribute('lighting-color') is "yellow"
PASS successfullyParsed is true

TEST COMPLETE
